tranaction-payment: break out tests & mock runtime to separate files
